+++ sample/src/androidMain/kotlin/com/github/panpf/sketch/sample/ui/test/transform/MaskTransformationTestFragment.kt
@@ -22,7 +22,7 @@ import androidx.core.graphics.ColorUtils
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Lifecycle.State
 import com.github.panpf.sketch.cache.CachePolicy.DISABLED
-import com.github.panpf.sketch.displayImage
+import com.github.panpf.sketch.loadImage
 import com.github.panpf.sketch.images.MyImages
 import com.github.panpf.sketch.sample.databinding.FragmentTestTransformationMaskBinding
 import com.github.panpf.sketch.sample.ui.base.BaseBindingFragment
@@ -47,7 +47,7 @@ class MaskTransformationTestFragment :
             binding.greenButton.isChecked = it == MaskTransformationTestViewModel.MaskColor.GREEN
             binding.blueButton.isChecked = it == MaskTransformationTestViewModel.MaskColor.BLUE
 
-            binding.myImage.displayImage(MyImages.png.uri) {
+            binding.myImage.loadImage(MyImages.png.uri) {
                 memoryCachePolicy(DISABLED)
                 resultCachePolicy(DISABLED)
                 addTransformations(MaskTransformation(it.colorInt))