+++ sample/src/androidMain/kotlin/com/github/panpf/sketch/sample/ui/test/transform/MultiTransformationTestFragment.kt
@@ -24,7 +24,7 @@ import androidx.core.graphics.ColorUtils
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Lifecycle.State
 import com.github.panpf.sketch.cache.CachePolicy.DISABLED
-import com.github.panpf.sketch.displayImage
+import com.github.panpf.sketch.loadImage
 import com.github.panpf.sketch.images.MyImages
 import com.github.panpf.sketch.sample.databinding.FragmentTestTransformationMultiBinding
 import com.github.panpf.sketch.sample.ui.base.BaseBindingFragment
@@ -135,7 +135,7 @@ class MultiTransformationTestFragment :
     }
 
     private fun updateImage(binding: FragmentTestTransformationMultiBinding) {
-        binding.myImage.displayImage(MyImages.statics.first().uri) {
+        binding.myImage.loadImage(MyImages.statics.first().uri) {
             memoryCachePolicy(DISABLED)
             resultCachePolicy(DISABLED)
             addTransformations(