+++ sample/src/androidMain/kotlin/com/github/panpf/sketch/sample/ui/test/transform/RoundedCornersTransformationTestFragment.kt
@@ -22,7 +22,7 @@ import android.widget.SeekBar
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Lifecycle.State
 import com.github.panpf.sketch.cache.CachePolicy.DISABLED
-import com.github.panpf.sketch.displayImage
+import com.github.panpf.sketch.loadImage
 import com.github.panpf.sketch.images.MyImages
 import com.github.panpf.sketch.sample.databinding.FragmentTestTransformationRoundedCornersBinding
 import com.github.panpf.sketch.sample.ui.base.BaseBindingFragment
@@ -138,7 +138,7 @@ class RoundedCornersTransformationTestFragment :
     }
 
     private fun updateImage(binding: FragmentTestTransformationRoundedCornersBinding) {
-        binding.myImage.displayImage(MyImages.statics.first().uri) {
+        binding.myImage.loadImage(MyImages.statics.first().uri) {
             memoryCachePolicy(DISABLED)
             resultCachePolicy(DISABLED)
             addTransformations(