+++ sample/src/androidMain/kotlin/com/github/panpf/sketch/sample/ui/widget/zoomimage/view/SketchZoomImageView.kt
@@ -35,7 +35,7 @@ import com.github.panpf.zoomimage.subsampling.ImageSource
  *
  * ```kotlin
  * val sketchZoomImageView = SketchZoomImageView(context)
- * sketchZoomImageView.displayImage("http://sample.com/sample.jpg") {
+ * sketchZoomImageView.loadImage("http://sample.com/sample.jpg") {
  *     placeholder(R.drawable.placeholder)
  *     crossfade()
  * }