+++ sketch-view-core/src/androidTest/kotlin/com/github/panpf/sketch/view/core/test/ImageViewExtensionsTest.kt
@@ -21,7 +21,7 @@ import com.github.panpf.sketch.disposeDisplay
 import com.github.panpf.sketch.imageResult
 import com.github.panpf.sketch.images.MyImages
 import com.github.panpf.sketch.request.ImageResult
-import com.github.panpf.sketch.test.singleton.displayImage
+import com.github.panpf.sketch.test.singleton.loadImage
 import com.github.panpf.sketch.test.utils.DelayTransformation
 import com.github.panpf.sketch.test.utils.MediumImageViewTestActivity
 import com.github.panpf.tools4a.test.ktx.getActivitySync
@@ -42,7 +42,7 @@ class ImageViewExtensionsTest {
 
         Assert.assertNull(imageView.drawable)
         runBlocking {
-            imageView.displayImage(MyImages.jpeg.uri).job.join()
+            imageView.loadImage(MyImages.jpeg.uri).job.join()
         }
         Assert.assertNotNull(imageView.drawable)
 
@@ -51,7 +51,7 @@ class ImageViewExtensionsTest {
         }
         Assert.assertNull(imageView.drawable)
         runBlocking {
-            imageView.displayImage(MyImages.png.uri) {
+            imageView.loadImage(MyImages.png.uri) {
                 resultCachePolicy(DISABLED)
                 memoryCachePolicy(DISABLED)
                 addTransformations(DelayTransformation {
@@ -70,17 +70,17 @@ class ImageViewExtensionsTest {
         Assert.assertNull(imageView.imageResult)
 
         runBlocking {
-            imageView.displayImage(MyImages.jpeg.uri).job.join()
+            imageView.loadImage(MyImages.jpeg.uri).job.join()
         }
         Assert.assertTrue(imageView.imageResult is ImageResult.Success)
 
         runBlocking {
-            imageView.displayImage("asset://fake.jpeg").job.join()
+            imageView.loadImage("asset://fake.jpeg").job.join()
         }
         Assert.assertTrue(imageView.imageResult is ImageResult.Error)
 
         runBlocking {
-            imageView.displayImage(MyImages.png.uri) {
+            imageView.loadImage(MyImages.png.uri) {
                 resultCachePolicy(DISABLED)
                 memoryCachePolicy(DISABLED)
                 addTransformations(DelayTransformation {