+++ sketch-view/src/main/kotlin/com/github/panpf/sketch/image_view_singleton_extensions.kt
@@ -30,8 +30,7 @@ import java.io.File
  *
  * You can set request params with a trailing lambda function [configBlock]
  */
-// TODO rename to loadImage
-fun ImageView.displayImage(
+fun ImageView.loadImage(
     uri: String?,
     configBlock: (ImageRequest.Builder.() -> Unit)? = null
 ): Disposable = context.sketch.enqueue(ImageRequest(this, uri, configBlock))
@@ -41,58 +40,58 @@ fun ImageView.displayImage(
  *
  * You can set request params with a trailing lambda function [configBlock]
  */
-fun ImageView.displayImage(
+fun ImageView.loadImage(
     uri: Uri?,
     configBlock: (ImageRequest.Builder.() -> Unit)? = null
-): Disposable = displayImage(uri?.toString(), configBlock)
+): Disposable = loadImage(uri?.toString(), configBlock)
 
 /**
  * Load the image from drawable res and display it on this [ImageView]
  *
  * You can set request params with a trailing lambda function [configBlock]
  */
-fun ImageView.displayImage(
+fun ImageView.loadImage(
     @DrawableRes drawableResId: Int?,
     configBlock: (ImageRequest.Builder.() -> Unit)? = null
-): Disposable = displayImage(drawableResId?.let { newResourceUri(it) }, configBlock)
+): Disposable = loadImage(drawableResId?.let { newResourceUri(it) }, configBlock)
 
 /**
  * Load the image from local file and display it on this [ImageView]
  *
  * You can set request params with a trailing lambda function [configBlock]
  */
-fun ImageView.displayImage(
+fun ImageView.loadImage(
     file: File?,
     configBlock: (ImageRequest.Builder.() -> Unit)? = null
-): Disposable = displayImage(file?.let { newFileUri(it.path) }, configBlock)
+): Disposable = loadImage(file?.let { newFileUri(it.path) }, configBlock)
 
 /**
  * Load the image from app assets and display it on this [ImageView]
  *
  * You can set request params with a trailing lambda function [configBlock]
  */
-fun ImageView.displayAssetImage(
+fun ImageView.loadAssetImage(
     assetFileName: String?,
     configBlock: (ImageRequest.Builder.() -> Unit)? = null
-): Disposable = displayImage(assetFileName?.let { newAssetUri(assetFileName) }, configBlock)
+): Disposable = loadImage(assetFileName?.let { newAssetUri(assetFileName) }, configBlock)
 
 /**
  * Load the image from drawable res and display it on this [ImageView]
  *
  * You can set request params with a trailing lambda function [configBlock]
  */
-fun ImageView.displayResourceImage(
+fun ImageView.loadResourceImage(
     @DrawableRes drawableResId: Int?,
     configBlock: (ImageRequest.Builder.() -> Unit)? = null
-): Disposable = displayImage(drawableResId?.let { newResourceUri(it) }, configBlock)
+): Disposable = loadImage(drawableResId?.let { newResourceUri(it) }, configBlock)
 
 /**
  * Load the image from drawable res and display it on this [ImageView]
  *
  * You can set request params with a trailing lambda function [configBlock]
  */
-fun ImageView.displayResourceImage(
+fun ImageView.loadResourceImage(
     packageName: String,
     @DrawableRes drawableResId: Int,
     configBlock: (ImageRequest.Builder.() -> Unit)? = null
-): Disposable = displayImage(newResourceUri(packageName, drawableResId), configBlock)
+): Disposable = loadImage(newResourceUri(packageName, drawableResId), configBlock)